added utils folder e datasets/edited folder and modified emfd_lemma analysis
